CONFIG-Checker/requirements.txt
requests
beautifulsoup4
schedule
python-dotenv
pyyaml